requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.10
billiard==3.3.0.23
Django==3.1
django-crontab==0.7.1
django-filter==2.3.0
djangorestframework==3.11.1
flake8==3.8.3
importlib-metadata==1.7.0
kombu==3.0.37
mccabe==0.6.1
psycopg2==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
